+++ InstaTistoryAutomatorServer/src/main/resources/application-test.yml
@@ -21,6 +21,6 @@ spring:
 
 webdriver:
   timeout: 3
-  chrome-driver-path: src/main/resources/chromedriver
+  chrome-driver-path: InstaTistoryAutomatorServer/src/main/resources/chromedriver
   arguments:
     - --lang=ko
