--- InstaTistoryAutomatorServer/src/test/kotlin/com/jun/instatistoryautomatorserver/YamlPropertySourceFactory.kt
@@ -1,18 +1,0 @@
-package com.jun.instatistoryautomatorserver
-
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean
-import org.springframework.core.env.PropertiesPropertySource
-import org.springframework.core.env.PropertySource
-import org.springframework.core.io.support.EncodedResource
-import org.springframework.core.io.support.PropertySourceFactory
-import java.io.IOException
-
-
-class YamlPropertySourceFactory : PropertySourceFactory {
-    @Throws(IOException::class)
-    override fun createPropertySource(name: String?, encodedResource: EncodedResource): PropertySource<*> {
-        val factory = YamlPropertiesFactoryBean()
-        factory.setResources(encodedResource.resource)
-        return PropertiesPropertySource(encodedResource.resource.filename!!, factory.getObject()!!)
-    }
-}
